feat: log memory usage in production too
