+++ src/index.ts
@@ -3,7 +3,8 @@ import { config } from "dotenv";
 import express from "express";
 import mongoose from "mongoose";
 import { join } from "node:path";
-import process, { cwd, env } from "node:process";
+import process, { cwd, env, memoryUsage } from "node:process";
+import { setInterval } from "node:timers/promises";
 import Constants, {
 	CustomClient,
 	loadPredictions,
@@ -62,3 +63,18 @@ for (const font in fonts)
 await mongoose.connect(env["MONGODB_URL"]!);
 await client.login();
 await loadPredictions(client);
+for await (const _ of setInterval(
+	60_000 * (env.NODE_ENV === "production" ? 10 : 1),
+)) {
+	const memory = memoryUsage();
+
+	printToStdout(
+		`RSS: ${(memory.rss / 1000 / 1000).toFixed(3)}MB\nHeap Used: ${(
+			memory.heapUsed /
+			1000 /
+			1000
+		).toFixed(3)}MB\nHeap Total: ${(memory.heapTotal / 1000 / 1000).toFixed(
+			3,
+		)}MB\nExternal: ${(memory.external / 1000 / 1000).toFixed(3)}MB`,
+	);
+}
